libADLMIDI now is buildable under OpenWatcom except of some workarounds - Need a different way to create a static library, otherwise it don't wanna be built. - Needed a different custom STL containers (std::vector and std::set) implementation that will work and will don't glitch with a crashes

+++ src/adlmidi.cpp
@@ -471,7 +471,7 @@ ADLMIDI_EXPORT const Adl_MarkerEntry adl_metaMarker(struct ADL_MIDIPlayer *devic
         MIDIplay::MIDI_MarkerEntry &mk = play->musMarkers[index];
         marker.label = mk.label.c_str();
         marker.pos_time = mk.pos_time;
-        marker.pos_ticks = mk.pos_ticks;
+        marker.pos_ticks = (unsigned long)mk.pos_ticks;
     }
     return marker;
 }
